requirements.txt
jieba>=0.42.1
snownlp>=0.12.3
pyyaml>=6.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.65.0
requests>=2.28.0
python-dotenv>=0.19.0